Get the collections implemented in the package

Get the collections implemented in the package and a set of related information

get_collections_available(collection)

Arguments

collection
string. Collection of interest.

Value

A data.frame with the available collections, and a set of related information. Columns are :

collection :Collections available through opendapr

source :source

long_name :Collection long name

url_metadata :URL pointing to metadata about the collection

DOI :DOI of the dataset

url_opendapserver :URL of the OPeNDAP server from which the data is extracted

status :Implemented status in sopendapr

crs :Coordinates reference system of the collection

Examples

(collections <- get_collections_available())

#>                 collection source
#> 1              MOD11A1.006  MODIS
#> 2              MYD11A1.006  MODIS
#> 3              MOD11A2.006  MODIS
#> 4              MYD11A2.006  MODIS
#> 5              MOD13Q1.006  MODIS
#> 6              MYD13Q1.006  MODIS
#> 7              MOD16A2.006  MODIS
#> 8              MYD16A2.006  MODIS
#> 9             VNP21A1D.001    VNP
#> 10            VNP21A1N.001    VNP
#> 11             VNP21A2.001    VNP
#> 12             VNP13A1.001    VNP
#> 13  GPM_L3/GPM_3IMERGDE.06    GPM
#> 14 GPM_L3/GPM_3IMERGHHE.06    GPM
#> 15 GPM_L3/GPM_3IMERGHHL.06    GPM
#> 16  GPM_L3/GPM_3IMERGDF.06    GPM
#> 17  GPM_L3/GPM_3IMERGHH.06    GPM
#> 18   GPM_L3/GPM_3IMERGM.06    GPM
#> 19  GPM_L3/GPM_3IMERGDL.06    GPM
#> 20      SMAP/SPL3SMP_E.003   SMAP
#> 21             SRTMGL1.003   SRTM
#>                                                                                     long_name
#> 1           MODIS/Terra Land Surface Temperature/Emissivity Daily L3 Global 1km SIN Grid V006
#> 2            MODIS/Aqua Land Surface Temperature/Emissivity Daily L3 Global 1km SIN Grid V006
#> 3          MODIS/Terra Land Surface Temperature/Emissivity 8-Day L3 Global 1 km SIN Grid V006
#> 4           MODIS/Aqua Land Surface Temperature/Emissivity 8-Day L3 Global 1 km SIN Grid V006
#> 5                          MODIS/Terra Vegetation Indices 16-Day L3 Global 250m SIN Grid V006
#> 6                           MODIS/Aqua Vegetation Indices 16-Day L3 Global 250m SIN Grid V006
#> 7                       MODIS/Terra Net Evapotranspiration 8-Day L4 Global 500m SIN Grid V006
#> 8                        MODIS/Aqua Net Evapotranspiration 8-Day L4 Global 500m SIN Grid V006
#> 9        VIIRS/NPP Land Surface Temperature and Emissivity Daily L3 Global 1 km SIN Grid Day 
#> 10     VIIRS/NPP Land Surface Temperature and Emissivity Daily L3 Global 1 km SIN Grid Night 
#> 11            VIIRS/NPP Land Surface Temperature and Emissivity 8-Day L3 Global 1 km SIN Grid
#> 12                               VIIRS/NPP Vegetation Indices 16-Day L3 Global 500 m SIN Grid
#> 13           GPM_3IMERGDE: GPM IMERG Early Precipitation L3 1 day 0.1 degree x 0.1 degree V06
#> 14    GPM_3IMERGHHE: GPM IMERG Early Precipitation L3 Half Hourly 0.1 degree x 0.1 degree V06
#> 15     GPM_3IMERGHHE: GPM IMERG Late Precipitation L3 Half Hourly 0.1 degree x 0.1 degree V06
#> 16                         GPM IMERG Final Precipitation L3 1 day 0.1 degree x 0.1 degree V06
#> 17                   GPM IMERG Final Precipitation L3 Half Hourly 0.1 degree x 0.1 degree V06
#> 18                       GPM IMERG Final Precipitation L3 1 month 0.1 degree x 0.1 degree V06
#> 19                          GPM IMERG Late Precipitation L3 1 day 0.1 degree x 0.1 degree V06
#> 20           SMAP Enhanced L3 Radiometer Global Daily 9 km EASE-Grid Soil Moisture, Version 3
#> 21 Digital Elevation Model from the NASA Shuttle Radar Topography Mission Global 1 arc second
#>                                                     url_metadata
#> 1                  https://lpdaac.usgs.gov/products/mod11a1v006/
#> 2                  https://lpdaac.usgs.gov/products/myd11a1v006/
#> 3                  https://lpdaac.usgs.gov/products/mod11a2v006/
#> 4                  https://lpdaac.usgs.gov/products/myd11a2v006/
#> 5                  https://lpdaac.usgs.gov/products/mod13q1v006/
#> 6                  https://lpdaac.usgs.gov/products/myd13q1v006/
#> 7                  https://lpdaac.usgs.gov/products/mod16a2v006/
#> 8                  https://lpdaac.usgs.gov/products/myd16a2v006/
#> 9                 https://lpdaac.usgs.gov/products/vnp21a1dv001/
#> 10                https://lpdaac.usgs.gov/products/vnp21a1nv001/
#> 11                 https://lpdaac.usgs.gov/products/vnp21a2v001/
#> 12                 https://lpdaac.usgs.gov/products/vnp13a1v001/
#> 13   https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGDE_06/summary
#> 14 https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGDHHE_06/summary
#> 15  https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGHHL_06/summary
#> 16   https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGDF_06/summary
#> 17   https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGHH_06/summary
#> 18    https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGM_06/summary
#> 19   https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGDL_06/summary
#> 20                           https://nsidc.org/data/spl3smp_e#sm
#> 21                 https://lpdaac.usgs.gov/products/srtmgl1v003/
#>                                                     DOI
#> 1          https://dx.doi.org/10.5067/MODIS/MOD11A1.006
#> 2          https://dx.doi.org/10.5067/MODIS/MYD11A1.006
#> 3          https://dx.doi.org/10.5067/MODIS/MOD11A2.006
#> 4          https://dx.doi.org/10.5067/MODIS/MYD11A2.006
#> 5          https://dx.doi.org/10.5067/MODIS/MOD13Q1.006
#> 6          https://dx.doi.org/10.5067/MODIS/MYD13Q1.006
#> 7          https://dx.doi.org/10.5067/MODIS/MOD16A2.006
#> 8          https://dx.doi.org/10.5067/MODIS/MYD16A2.006
#> 9            https://doi.org/10.5067/VIIRS/VNP21A1D.001
#> 10           https://doi.org/10.5067/VIIRS/VNP21A1N.001
#> 11            https://doi.org/10.5067/VIIRS/VNP21A2.001
#> 12            https://doi.org/10.5067/VIIRS/VNP13A1.001
#> 13           https://doi.org/10.5067/GPM/IMERGDE/DAY/06
#> 14         https://doi.org/10.5067/GPM/IMERG/3B-HH-E/06
#> 15        https://doi.org/10.5067/GPM/IMERG/3B-HH-L/06 
#> 16           https://doi.org/10.5067/GPM/IMERGDF/DAY/06
#> 17           https://doi.org/10.5067/GPM/IMERG/3B-HH/06
#> 18        https://doi.org/10.5067/GPM/IMERG/3B-MONTH/06
#> 19           https://doi.org/10.5067/GPM/IMERGDL/DAY/06
#> 20                 https://doi.org/10.5067/T90W6VRLCBHI
#> 21 https://dx.doi.org/10.5067/MEASURES/SRTM/SRTMGL1.003
#>                                                        url_opendapserver
#> 1    https://opendap.cr.usgs.gov/opendap/hyrax/MOD11A1.006/contents.html
#> 2    https://opendap.cr.usgs.gov/opendap/hyrax/MYD11A1.006/contents.html
#> 3    https://opendap.cr.usgs.gov/opendap/hyrax/MOD11A2.006/contents.html
#> 4    https://opendap.cr.usgs.gov/opendap/hyrax/MYD11A2.006/contents.html
#> 5    https://opendap.cr.usgs.gov/opendap/hyrax/MOD13Q1.006/contents.html
#> 6    https://opendap.cr.usgs.gov/opendap/hyrax/MYD13Q1.006/contents.html
#> 7    https://opendap.cr.usgs.gov/opendap/hyrax/MOD16A2.006/contents.html
#> 8    https://opendap.cr.usgs.gov/opendap/hyrax/MYD16A2.006/contents.html
#> 9   https://opendap.cr.usgs.gov/opendap/hyrax/VNP21A1D.001/contents.html
#> 10  https://opendap.cr.usgs.gov/opendap/hyrax/VNP21A1N.001/contents.html
#> 11   https://opendap.cr.usgs.gov/opendap/hyrax/VNP21A2.001/contents.html
#> 12   https://opendap.cr.usgs.gov/opendap/hyrax/VNP13A1.001/contents.html
#> 13  https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGDE.06/
#> 14 https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHHE.06/
#> 15 https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHHL.06/
#> 16  https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGDF.06/
#> 17  https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHH.06/
#> 18   https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGM.06/
#> 19  https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGDL.06/
#> 20            https://n5eil02u.ecs.nsidc.org/opendap/SMAP/SPL3SMP_E.003/
#> 21   https://opendap.cr.usgs.gov/opendap/hyrax/SRTMGL1.003/contents.html
#>                                                                                                                        crs
#> 1                                        +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> 2                                        +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> 3                                        +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> 4                                        +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> 5                                        +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> 6                                        +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> 7                                        +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> 8                                        +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> 9                                        +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> 10                                       +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> 11                                       +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> 12                                       +proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs
#> 13                                         +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> 14                                         +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,1
#> 15                                         +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,2
#> 16                                         +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,3
#> 17                                         +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,4
#> 18                                         +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,5
#> 19                                         +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> 20 +init=epsg:6933 +proj=cea +lon_0=0 +lat_ts=30 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0 
#> 21                                         +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#>         status
#> 1  Implemented
#> 2  Implemented
#> 3  Implemented
#> 4  Implemented
#> 5  Implemented
#> 6  Implemented
#> 7  Implemented
#> 8  Implemented
#> 9  Implemented
#> 10 Implemented
#> 11 Implemented
#> 12 Implemented
#> 13 Implemented
#> 14 Implemented
#> 15 Implemented
#> 16 Implemented
#> 17 Implemented
#> 18 Implemented
#> 19 Implemented
#> 20 Implemented
#> 21 Implemented