Get the URL of

some descirption

get_url(collection, variables = NULL, roi, time_range,
  output_format = "nc4", single_netcdf = TRUE, opt_param = NULL,
  credentials = NULL, verbose = FALSE)

Arguments

collection
string. Collection of interest.

variables
string vector. Variables to retrieve for the collection of interest. If not specified (default) all available variables will be extracted.

roi
object of class sf or sfc Region of interest. Must be POLYGON-type geometry. Can be composed of several features (see details).

time_range
date(s) / POSIXlt of interest (single date/datetime or time frame : vector with start and end dates/times) (see details)

output_format
string. Output format. Available options are : "nc4" (default), "ascii", "json"

single_netcdf
boolean. Get the URL either as a single file that encompasses the whole time frame (TRUE) or as multiple files (1 for each date) (FALSE). Default to TRUE. Currently enabled only for MODIS and VIIRS collections.

opt_param
list of optional arguments (see details). This parameter is the output of the function get_optional_parameters.

credentials
vector string. In case of data that needs login : string vector of length 2 with username and password

verbose
boolean. Verbose (default FALSE)

Value

a data.frame with one row for each dataset to download and 4 columns :

*time_start* : Start Date/time for the dataset

*name* : An indicative name for the dataset

*url* : http URL (OPeNDAP) of the dataset

*destfile* : An indicative destination file for the dataset

Details

This is the main function of the package. It enables to retrieve

#' Argument time_range can be provided either as a single date (e.g. as.Date("2017-01-01")) or time frame provided as two bounding dates ( e.g. as.Date(c("2010-01-01","2010-01-30"))) or as a POSIXlt single time or time range (e.g. "2010-01-01 18:00:00") for the half-hourly collection (GPM_3IMERGHH.06). If POSIXlt, times must be in UTC.

Argument optionals_opendap is optional. This parameter is automatically calculated within the function if it is not provided. However, providing it optimizes the performances of the function (i.e. fasten the processing time). It might be particularly useful to provide it if looping over the same ROI or dates is planned. The parameter can be retrieved outside the function with the function get_optional_parameters.

Note

some note

Examples

### First login to Earthdata
log <- login(c(Sys.getenv("earthdata_un"),Sys.getenv("earthdata_pw")))

#> Checking credentials...

#> Error in login(c(Sys.getenv("earthdata_un"), Sys.getenv("earthdata_pw"))): l'argument "source" est manquant, avec aucune valeur par défaut

############################################################
### Retrieve the URLs (OPeNDAP) to download the following datasets :
# MODIS Terra LST Daily (MOD11A1.006) (collection)
# Day + Night bands (LST_Day_1km,LST_Night_1km) (variables)
# over a 50km x 70km region of interest (roi)
# for the time frame 2017-01-01 to 2017-01-30 (30 days) (time_range)

(opendap_urls_mod11a1 <- get_url(collection = "MOD11A1.006",
variables = c("LST_Day_1km","LST_Night_1km"),
roi = sf::st_read(system.file("extdata/roi_example.gpkg", package = "opendapr"),quiet=TRUE),
time_range = as.Date(c("2017-01-01","2017-01-30"))
))

#> Error in login(credentials): credentials must be a vector character string of length 2 (username and password)

############################################################

### Download the data :

res_dl <- download_data(opendap_urls_mod11a1)

#> Error in inherits(df_to_dl, "data.frame"): objet 'opendap_urls_mod11a1' introuvable

### Open the data :
# When opening the data, do not forget to reset the CRS
# here, MODIS crs : https://spatialreference.org/ref/sr-org/modis-sinusoidal-3/proj4/
modis_crs <- "+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs"

## open as a stars object (deals with multiple dimensions and time)

require(stars)

#> Le chargement a nécessité le package : stars

#> Le chargement a nécessité le package : abind

require(magrittr)

#> Le chargement a nécessité le package : magrittr

#> 
#> Attachement du package : ‘magrittr’

#> The following objects are masked from ‘package:testthat’:
#> 
#>     equals, is_less_than, not

(mod11a1_stars <- read_stars(res_dl$destfile) %>% st_set_crs(modis_crs))

#> Error in read_stars(res_dl$destfile): objet 'res_dl' introuvable

plot(mod11a1_stars)

#> Error in plot(mod11a1_stars): objet 'mod11a1_stars' introuvable

## or open as a rasterBrick object (1 rasterBrick for each dimension)

require(raster)

#> Le chargement a nécessité le package : raster

#> Le chargement a nécessité le package : sp

#> 
#> Attachement du package : ‘raster’

#> The following object is masked from ‘package:magrittr’:
#> 
#>     extract

(mod11a1_rast_day <- brick(res_dl$destfile,varname="LST_Day_1km",crs=modis_crs))

#> Error in brick(res_dl$destfile, varname = "LST_Day_1km", crs = modis_crs): objet 'res_dl' introuvable

(mod11a1_rast_night <- brick(res_dl$destfile,varname="LST_Night_1km",crs=modis_crs))

#> Error in brick(res_dl$destfile, varname = "LST_Night_1km", crs = modis_crs): objet 'res_dl' introuvable

plot(mod11a1_rast_day[[1]])

#> Error in plot(mod11a1_rast_day[[1]]): objet 'mod11a1_rast_day' introuvable

# Here the time dimension is set as the z dimension :
(getZ(mod11a1_rast_day))

#> Error in getZ(mod11a1_rast_day): objet 'mod11a1_rast_day' introuvable

# Check out the vignettes for additional examples and more complex data import workflows