Get informations related to the variables available for a given collection

Get the variables available for a given collection

get_variables_info(collection, credentials = NULL)

Arguments

collection
string. mandatory. Collection of interest (see details of get_url).

credentials
vector string of length 2 with username and password. optional.

Value

A data.frame with the available variables for the collection, and a set of related information for each variable. The variables marked as "extractable" in the column "extractable_w_opendapr" can be provided as input parameter variables of the function get_url

Examples

# login to Earthdata
log <- login(c(Sys.getenv("earthdata_un"),Sys.getenv("earthdata_pw")),source="earthdata")

#> Checking credentials...
#> Successfull login to earthdata 

# Get the variables available for the collection MOD11A1.006
(df_varinfo <- get_variables_info("MOD11A1.006"))

#>                                          name
#> 1                                    Latitude
#> 2                                   Longitude
#> 3                                        YDim
#> 4                                        XDim
#> 5  MODIS_Grid_Daily_1km_LST_eos_cf_projection
#> 6                             Night_view_time
#> 7                             Clear_night_cov
#> 8                               LST_Night_1km
#> 9                                      QC_Day
#> 10                                   QC_Night
#> 11                              Day_view_time
#> 12                            Night_view_angl
#> 13                              Clear_day_cov
#> 14                              Day_view_angl
#> 15                                LST_Day_1km
#> 16                                    Emis_31
#> 17                                    Emis_32
#> 18                                       time
#>                                                       long_name units
#> 1                                                      Latitude  <NA>
#> 2                                                     Longitude  <NA>
#> 3  y coordinate of projection for grid MODIS_Grid_Daily_1km_LST meter
#> 4  x coordinate of projection for grid MODIS_Grid_Daily_1km_LST meter
#> 5                                                          <NA>  <NA>
#> 6        Time of nighttime Land-surface Temperature observation   hrs
#> 7                                      night clear-sky coverage  <NA>
#> 8             Daily nighttime 1km grid Land-surface Temperature     K
#> 9                Quality control for daytime LST and emissivity  <NA>
#> 10             Quality control for nighttime LST and emissivity  <NA>
#> 11         Time of daytime Land-surface Temperature observation   hrs
#> 12      View zenith angle of nighttime Land-surface Temperature   deg
#> 13                                       day clear-sky coverage  <NA>
#> 14        View zenith angle of daytime Land-surface Temperature   deg
#> 15              Daily daytime 1km grid Land-surface Temperature     K
#> 16                                           Band 31 emissivity  <NA>
#> 17                                           Band 32 emissivity  <NA>
#> 18                                                         <NA>  <NA>
#>                                      indices
#> 1                [YDim = 1200] [XDim = 1200]
#> 2                [YDim = 1200] [XDim = 1200]
#> 3                              [YDim = 1200]
#> 4                              [XDim = 1200]
#> 5                                       <NA>
#> 6  [time = 6841] [YDim = 1200] [XDim = 1200]
#> 7  [time = 6841] [YDim = 1200] [XDim = 1200]
#> 8  [time = 6841] [YDim = 1200] [XDim = 1200]
#> 9  [time = 6841] [YDim = 1200] [XDim = 1200]
#> 10 [time = 6841] [YDim = 1200] [XDim = 1200]
#> 11 [time = 6841] [YDim = 1200] [XDim = 1200]
#> 12 [time = 6841] [YDim = 1200] [XDim = 1200]
#> 13 [time = 6841] [YDim = 1200] [XDim = 1200]
#> 14 [time = 6841] [YDim = 1200] [XDim = 1200]
#> 15 [time = 6841] [YDim = 1200] [XDim = 1200]
#> 16 [time = 6841] [YDim = 1200] [XDim = 1200]
#> 17 [time = 6841] [YDim = 1200] [XDim = 1200]
#> 18                             [time = 6841]
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        all_info
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                               Array of 64 bit Reals [YDim = 0..1199][XDim = 0..1199]long_name: Latitude\nunits: degrees_north
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                               Array of 64 bit Reals [YDim = 0..1199][XDim = 0..1199]long_name: Longitude\nunits: degrees_east
#> 3                                                                                                                                                                                                                                                                                                                                                                                Array of 64 bit Reals [YDim = 0..1199]standard_name: projection_y_coordinate\nlong_name: y coordinate of projection for grid MODIS_Grid_Daily_1km_LST\nunits: meter\n_CoordinateAxisType: GeoY
#> 4                                                                                                                                                                                                                                                                                                                                                                                Array of 64 bit Reals [XDim = 0..1199]standard_name: projection_x_coordinate\nlong_name: x coordinate of projection for grid MODIS_Grid_Daily_1km_LST\nunits: meter\n_CoordinateAxisType: GeoX
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                             Bytegrid_mapping_name: sinusoidal\nlongitude_of_central_meridian: 0.0\nearth_radius: 6371007.181\n_CoordinateAxisTypes: GeoX GeoY
#> 6                                                        Array of 32 bit Reals [time = 0..6840][YDim = 0..1199][XDim = 0..1199]coordinates: Latitude Longitude\nlong_name: Time of nighttime Land-surface Temperature observation\nunits: hrs\nView_time: View_time data * scale_factor\nscale_factor_err: 0.0000000000000000\nadd_offset_err: 0.0000000000000000\ncalibrated_nt: 5\norig_scale_factor: 0.10000000000000001\n_FillValue: 255\nvalid_min: 0.000000000\nvalid_max: 24.00000000\nNumber_Type_Orig: uint8\ngrid_mapping: MODIS_Grid_Daily_1km_LST_eos_cf_projection
#> 7                                                                          Array of 32 bit Reals [time = 0..6840][YDim = 0..1199][XDim = 0..1199]coordinates: Latitude Longitude\nlong_name: night clear-sky coverage\nnight_clear_sky_cov: nit_clear_sky_cov data * scale_factor\nscale_factor_err: 0.0000000000000000\nadd_offset_err: 0.0000000000000000\ncalibrated_nt: 5\norig_scale_factor: 0.00050000000000000001\n_FillValue: 0\nvalid_min: 0.0005000000237\nvalid_max: 32.76750183\nNumber_Type_Orig: uint16\ngrid_mapping: MODIS_Grid_Daily_1km_LST_eos_cf_projection
#> 8                                                                           Array of 32 bit Reals [time = 0..6840][YDim = 0..1199][XDim = 0..1199]coordinates: Latitude Longitude\nlong_name: Daily nighttime 1km grid Land-surface Temperature\nunits: K\nLST: LST data * scale_factor\nscale_factor_err: 0.0000000000000000\nadd_offset_err: 0.0000000000000000\ncalibrated_nt: 5\norig_scale_factor: 0.020000000000000000\n_FillValue: 0\nvalid_min: 150.0000000\nvalid_max: 1310.699951\nNumber_Type_Orig: uint16\ngrid_mapping: MODIS_Grid_Daily_1km_LST_eos_cf_projection
#> 9                                                                                                                                                                                                                                                                                                                  Array of Bytes [time = 0..6840][YDim = 0..1199][XDim = 0..1199]coordinates: Latitude Longitude\nlong_name: Quality control for daytime LST and emissivity\nNumber_Type: uint8\nvalid_range: 0, 255\ngrid_mapping: MODIS_Grid_Daily_1km_LST_eos_cf_projection
#> 10                                                                                                                                                                                                                                                                                                               Array of Bytes [time = 0..6840][YDim = 0..1199][XDim = 0..1199]coordinates: Latitude Longitude\nlong_name: Quality control for nighttime LST and emissivity\nNumber_Type: uint8\nvalid_range: 0, 255\ngrid_mapping: MODIS_Grid_Daily_1km_LST_eos_cf_projection
#> 11                                                         Array of 32 bit Reals [time = 0..6840][YDim = 0..1199][XDim = 0..1199]coordinates: Latitude Longitude\nlong_name: Time of daytime Land-surface Temperature observation\nunits: hrs\nView_time: View_time data * scale_factor\nscale_factor_err: 0.0000000000000000\nadd_offset_err: 0.0000000000000000\ncalibrated_nt: 5\norig_scale_factor: 0.10000000000000001\n_FillValue: 255\nvalid_min: 0.000000000\nvalid_max: 24.00000000\nNumber_Type_Orig: uint8\ngrid_mapping: MODIS_Grid_Daily_1km_LST_eos_cf_projection
#> 12 Array of 32 bit Reals [time = 0..6840][YDim = 0..1199][XDim = 0..1199]coordinates: Latitude Longitude\nlong_name: View zenith angle of nighttime Land-surface Temperature\nunits: deg\nView_angle: View_angle data * scale_factor + add_offset\nscale_factor_err: 0.0000000000000000\nadd_offset_err: 0.0000000000000000\ncalibrated_nt: 5\norig_scale_factor: 1.0000000000000000\norig_add_offset: -65.000000000000000\n_FillValue: 255\nvalid_min: -65.00000000\nvalid_max: 65.00000000\nNumber_Type_Orig: uint8\ngrid_mapping: MODIS_Grid_Daily_1km_LST_eos_cf_projection
#> 13                                                                             Array of 32 bit Reals [time = 0..6840][YDim = 0..1199][XDim = 0..1199]coordinates: Latitude Longitude\nlong_name: day clear-sky coverage\nday_clear_sky_cov: day_clear_sky_cov data * scale_factor\nscale_factor_err: 0.0000000000000000\nadd_offset_err: 0.0000000000000000\ncalibrated_nt: 5\norig_scale_factor: 0.00050000000000000001\n_FillValue: 0\nvalid_min: 0.0005000000237\nvalid_max: 32.76750183\nNumber_Type_Orig: uint16\ngrid_mapping: MODIS_Grid_Daily_1km_LST_eos_cf_projection
#> 14   Array of 32 bit Reals [time = 0..6840][YDim = 0..1199][XDim = 0..1199]coordinates: Latitude Longitude\nlong_name: View zenith angle of daytime Land-surface Temperature\nunits: deg\nView_angle: View_angle data * scale_factor + add_offset\nscale_factor_err: 0.0000000000000000\nadd_offset_err: 0.0000000000000000\ncalibrated_nt: 5\norig_scale_factor: 1.0000000000000000\norig_add_offset: -65.000000000000000\n_FillValue: 255\nvalid_min: -65.00000000\nvalid_max: 65.00000000\nNumber_Type_Orig: uint8\ngrid_mapping: MODIS_Grid_Daily_1km_LST_eos_cf_projection
#> 15                                                                            Array of 32 bit Reals [time = 0..6840][YDim = 0..1199][XDim = 0..1199]coordinates: Latitude Longitude\nlong_name: Daily daytime 1km grid Land-surface Temperature\nunits: K\nLST: LST data * scale_factor\nscale_factor_err: 0.0000000000000000\nadd_offset_err: 0.0000000000000000\ncalibrated_nt: 5\norig_scale_factor: 0.020000000000000000\n_FillValue: 0\nvalid_min: 150.0000000\nvalid_max: 1310.699951\nNumber_Type_Orig: uint16\ngrid_mapping: MODIS_Grid_Daily_1km_LST_eos_cf_projection
#> 16                                                       Array of 32 bit Reals [time = 0..6840][YDim = 0..1199][XDim = 0..1199]coordinates: Latitude Longitude\nlong_name: Band 31 emissivity\nEmis_31: Emis_31 data * scale_factor + add_offset\nscale_factor_err: 0.0000000000000000\nadd_offset_err: 0.0000000000000000\ncalibrated_nt: 5\norig_scale_factor: 0.0020000000000000000\norig_add_offset: 0.48999999999999999\n_FillValue: 0\nvalid_min: 0.4920000136\nvalid_max: 1.000000000\nNumber_Type_Orig: uint8\ngrid_mapping: MODIS_Grid_Daily_1km_LST_eos_cf_projection
#> 17                                                       Array of 32 bit Reals [time = 0..6840][YDim = 0..1199][XDim = 0..1199]coordinates: Latitude Longitude\nlong_name: Band 32 emissivity\nEmis_32: Emis_32 data * scale_factor + add_offset\nscale_factor_err: 0.0000000000000000\nadd_offset_err: 0.0000000000000000\ncalibrated_nt: 5\norig_scale_factor: 0.0020000000000000000\norig_add_offset: 0.48999999999999999\n_FillValue: 0\nvalid_min: 0.4920000136\nvalid_max: 1.000000000\nNumber_Type_Orig: uint8\ngrid_mapping: MODIS_Grid_Daily_1km_LST_eos_cf_projection
#> 18                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Array of 64 bit Reals [time = 0..6840]units: days since 2000-01-01 00:00
#>     extractable_w_opendapr
#> 1          not extractable
#> 2          not extractable
#> 3  automatically extracted
#> 4  automatically extracted
#> 5  automatically extracted
#> 6              extractable
#> 7              extractable
#> 8              extractable
#> 9              extractable
#> 10             extractable
#> 11             extractable
#> 12             extractable
#> 13             extractable
#> 14             extractable
#> 15             extractable
#> 16             extractable
#> 17             extractable
#> 18 automatically extracted