Get the collections implemented in the package

Get the collections implemented in the package and a set of related information

odr_list_collections()

Value

A data.frame with the available collections, and a set of related information. Main columns are :

collection

Collections name

source

Data provider

long_name

Collection long name

doi

DOI of the collection

start_date

Start date for the collection

url_opendapserver

URL of the OPeNDAP server from which the data is extracted

Examples

(head(odr_list_collections()))

#>         collection source
#> 1  GPM_3IMERGDE.06    GPM
#> 2  GPM_3IMERGDF.06    GPM
#> 3  GPM_3IMERGDL.06    GPM
#> 4  GPM_3IMERGHH.06    GPM
#> 5 GPM_3IMERGHHE.06    GPM
#> 6 GPM_3IMERGHHL.06    GPM
#>                                                                  long_name
#> 1       GPM IMERG Early Precipitation L3 1 day 0.1 degree x 0.1 degree V06
#> 2       GPM IMERG Final Precipitation L3 1 day 0.1 degree x 0.1 degree V06
#> 3        GPM IMERG Late Precipitation L3 1 day 0.1 degree x 0.1 degree V06
#> 4 GPM IMERG Final Precipitation L3 Half Hourly 0.1 degree x 0.1 degree V06
#> 5 GPM IMERG Early Precipitation L3 Half Hourly 0.1 degree x 0.1 degree V06
#> 6  GPM IMERG Late Precipitation L3 Half Hourly 0.1 degree x 0.1 degree V06
#>     nature      provider
#> 1 Rainfall NASA GES DISC
#> 2 Rainfall NASA GES DISC
#> 3 Rainfall NASA GES DISC
#> 4 Rainfall NASA GES DISC
#> 5 Rainfall NASA GES DISC
#> 6 Rainfall NASA GES DISC
#>                                                    url_metadata
#> 1   https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGDE_06/summary
#> 2   https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGDF_06/summary
#> 3   https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGDL_06/summary
#> 4   https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGHH_06/summary
#> 5 https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGDHHE_06/summary
#> 6  https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGHHL_06/summary
#>                                             doi version spatial_resolution_m
#> 1    https://doi.org/10.5067/GPM/IMERGDE/DAY/06       6                10000
#> 2    https://doi.org/10.5067/GPM/IMERGDF/DAY/06       6                10000
#> 3    https://doi.org/10.5067/GPM/IMERGDL/DAY/06       6                10000
#> 4    https://doi.org/10.5067/GPM/IMERG/3B-HH/06       6                10000
#> 5  https://doi.org/10.5067/GPM/IMERG/3B-HH-E/06       6                10000
#> 6 https://doi.org/10.5067/GPM/IMERG/3B-HH-L/06        6                10000
#>   temporal_resolution temporal_resolution_unit spatial_coverage start_date
#> 1                   1                      day           Global 2000-06-01
#> 2                   1                      day           Global 2000-06-01
#> 3                   1                      day           Global 2000-06-01
#> 4                  30                   minute           Global 2000-06-01
#> 5                  30                   minute           Global 2000-06-01
#> 6                  30                   minute           Global 2000-06-01
#>   end_date indicative_latency_days
#> 1  ongoing                       1
#> 2  ongoing                     100
#> 3  ongoing                       1
#> 4  ongoing                     100
#> 5  ongoing                       0
#> 6  ongoing                       1
#>                                             url_manual_access      status
#> 1  https://search.earthdata.nasa.gov/search?q=GPM_3IMERGDE_06 Implemented
#> 2  https://search.earthdata.nasa.gov/search?q=GPM_3IMERGDF_06 Implemented
#> 3  https://search.earthdata.nasa.gov/search?q=GPM_3IMERGDL_06 Implemented
#> 4  https://search.earthdata.nasa.gov/search?q=GPM_3IMERGHH_06 Implemented
#> 5 https://search.earthdata.nasa.gov/search?q=GPM_3IMERGHHE_06 Implemented
#> 6 https://search.earthdata.nasa.gov/search?q=GPM_3IMERGHHL_06 Implemented
#>       login                                    url_opendapserver
#> 1 earthdata https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/
#> 2 earthdata https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/
#> 3 earthdata https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/
#> 4 earthdata https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/
#> 5 earthdata https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/
#> 6 earthdata https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/
#>                                                 url_programmatic_access
#> 1  https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGDE.06/
#> 2  https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGDF.06/
#> 3  https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGDL.06/
#> 4  https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHH.06/
#> 5 https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHHE.06/
#> 6 https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHHL.06/
#>                                                                                                                                url_opendapexample
#> 1             https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGDE.06/2000/06/3B-DAY-E.MS.MRG.3IMERG.20000601-S000000-E235959.V06.nc4
#> 2         https://gpm1.gesdisc.eosdis.nasa.gov/opendap/hyrax/GPM_L3/GPM_3IMERGDF.06/2000/06/3B-DAY.MS.MRG.3IMERG.20000601-S000000-E235959.V06.nc4
#> 3       https://gpm1.gesdisc.eosdis.nasa.gov/opendap/hyrax/GPM_L3/GPM_3IMERGDL.06/2000/06/3B-DAY-L.MS.MRG.3IMERG.20000601-S000000-E235959.V06.nc4
#> 4 https://gpm1.gesdisc.eosdis.nasa.gov/opendap/hyrax/GPM_L3/GPM_3IMERGHH.06/2000/153/3B-HHR.MS.MRG.3IMERG.20000601-S000000-E002959.0000.V06B.HDF5
#> 5    https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHHE.06/2000/153/3B-HHR-E.MS.MRG.3IMERG.20000601-S000000-E002959.0000.V06B.HDF5
#> 6    https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHHL.06/2000/153/3B-HHR-L.MS.MRG.3IMERG.20000601-S000000-E002959.0000.V06B.HDF5
#>   dim_lon dim_lat dim_time dim_proj
#> 1     lon     lat     time     <NA>
#> 2     lon     lat     time     <NA>
#> 3     lon     lat     time     <NA>
#> 4     lon     lat     time     <NA>
#> 5     lon     lat     time     <NA>
#> 6     lon     lat     time     <NA>
#>                                                                               crs
#> 1 +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> 2 +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,3
#> 3 +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> 4 +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,4
#> 5 +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,1
#> 6 +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,2